Removing lru cache dependency and updating deb package
