+++ .travis/setup_kong.sh
@@ -9,7 +9,7 @@ sudo apt-get install wget curl tar make gcc unzip git liblua5.1-0-dev
 sudo apt-get install sudo netcat lua5.1 openssl libpcre3 dnsmasq
 
 # Installing Kong and its dependencies
-sudo dpkg -i ./.travis/kong-0.3.0travis.precise_all.deb
+sudo dpkg -i ./.travis/kong-0.4.0travis.precise_all.deb
 
 sudo luarocks remove kong --force
 sudo rm -rf /etc/kong
