--- kong/plugins/ip-restriction/init_worker.lua
@@ -1,9 +1,0 @@
-local iputils = require "resty.iputils"
-
-local _M = {}
-
-function _M.execute()
-  iputils.enable_lrucache()
-end
-
-return _M
