+++ kong/plugins/ip_restriction/init_worker.lua
@@ -1,0 +1,9 @@
+local iputils = require "resty.iputils"
+
+local _M = {}
+
+function _M.execute()
+  iputils.enable_lrucache()
+end
+
+return _M
